requirements.txt
requests
pandas
bs4
python-dotenv